src/main/java/net/transformatorhuis/cgi/elements/primitives/RiSphere.java
package net.transformatorhuis.cgi.elements.primitives;

import org.apache.log4j.Logger;
import net.transformatorhuis.cgi.conversion.AbstractRibType;
import net.transformatorhuis.cgi.conversion.AbstractRib;

/**
 * @author cyberroadie
 *
 * RISpec 3.2 page 78
 */
public class RiSphere extends AbstractRib {

    /**
     * Logger.
     */
    private static Logger logger = Logger.getLogger(RiSphere.class);

    /**
     * Sphere.
     */
    public RiSphere() {
        super();
    }

    /**
     * Sphere.
     * @param param extra paramters
     */
    public RiSphere(final String param) {
        super(param);
    }

    /**
     * Abstract method to return
     *
     * @return JAXB node returned for RIB element
     */
    public Object getJAXBNode() {
        return null;
    }

}
